Add mobile emulation support: implement GBrowserMobileToggleAction with device toolbar, rotation controls, responsive mode, and improved scaling logic; extend tests for state management, utilities, and edge cases.
